Print out characters in final solution
This is a little tricky in that if multiple dictionary
words match a jumbled word, you can end up with more than one
set of circled, unjumbled characters. This tries to represent
the alternatives. The point here is to allow the human to see
what characters could fit in the final answer.